requirements.txt
anyio==4.9.0
asgiref==3.8.1
astroid==3.3.11
click==8.1.8
colorama==0.4.6
coverage==7.8.0
dill==0.4.0
Django==5.1.4
django-cors-headers==4.6.0
django-rest-passwordreset==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httptools==0.6.4
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.8
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
pyotp==2.9.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.3
tomlkit==0.13.3
tzdata==2024.2
uritemplate==4.1.1
watchfiles==1.0.4
websockets==15.0.1
